refactor(table): reorganize filters module and unify membership
Splits filters.ts into a filters/ folder by concern (types, operators, guards, builders, defaults, serialize, format). Removes the 'select' filter type; text and number columns now carry their own in/not_in membership shape. Consolidates operator groups so OPERATORS_WITHOUT_VALUE and per-type op lists derive from one source, and promotes supportsMembership to a type guard (isMembershipFilterType) used at all call sites. columnEditableType now throws on invalid filterType instead of silently falling back to text, and is reused in handleColumnChange. Drops the trivial getEditableType wrapper and the ColumnFilterValue cast in defaultFilterValueFor's membership branch. Renames renderSortFilterIcon to renderSortIcon and narrows filterType once in context-menu. Editor defaults for date/datetime are now '=='.
